# cloud/backend/mon-projet-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel 2026",
        "description": "Documentation de mon API générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Lister les status",
                "operationId": "8bc31408b679bd52a317923523467108",
                "responses": {
                    "200": {
                        "description": "Liste des status"
                    }
                }
            },
            "post": {
                "tags": [
                    "Status"
                ],
                "summary": "Créer un status",
                "operationId": "9feef4e3b031caedf53a634533847d40",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "niveau"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Nouveau"
                                    },
                                    "niveau": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_deleted": {
                                        "type": "string",
                                        "example": "non"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Status créé"
                    }
                }
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Afficher un status",
                "operationId": "700d24ee8a044d063a878f2c4555e35f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Status"
                ],
                "summary": "Modifier un status",
                "operationId": "9bb972cfd17ab981f264046328e07e8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Status"
                ],
                "summary": "Supprimer un status",
                "operationId": "7bf136ee555a4babc4b933d4536adf2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status supprimé"
                    }
                }
            }
        },
...
        "/api/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister les utilisateurs",
                "operationId": "6e23f57bbb8711de48e707b59792258a",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "2f9bff660d3ec2acb423971a29aec9f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mdp",
                                    "Id_role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "Id_role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "fire_user_id": {
                                        "type": "string",
                                        "example": "b7Hk6pV0bVQq..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    }
                }
            }
        },
        "/api/utilisateurs/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher un utilisateur",
                "operationId": "ddaf85fef536a4e67b058c82124effbb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Modifier un utilisateur",
                "operationId": "d3071c3380dc44fe0061e7f6019eae98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "7f2e091a07c9b4ad7874d65acdb12ace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    }
                }
            }
        }
    },
...
}